+++ backend/errors/ListingError.js
@@ -1,0 +1,9 @@
+class ListingError extends Error {
+  constructor(message) {
+    super(message);
+
+    this.name = 'ListingError';
+  }
+}
+
+module.exports = ListingError;
