+++ backend/errors/index.js
@@ -1,0 +1,9 @@
+const MongoDbError = require('./MongoDbError');
+const SchemaValidationError = require('./SchemaValidationError');
+const ListingError = require('./ListingError');
+
+module.exports = {
+  MongoDbError,
+  SchemaValidationError,
+  ListingError,
+};
