+++ backend/joiSchemas/index.js
@@ -1,0 +1,7 @@
+const createListingSchema = require('./createListingSchema');
+const retrieveListingSchema = require('./retrieveListingSchema');
+
+module.exports = {
+  createListingSchema,
+  retrieveListingSchema,
+};
