+++ backend/lib/index.js
@@ -1,11 +1,5 @@
-const MongoDbError = require('./MongoDbError');
-const listingJoiSchema = require('./listingJoiSchema');
 const handleJoiValidationError = require('./handleJoiValidationError');
-const SchemaValidationError = require('./SchemaValidationError');
 
 module.exports = {
-  MongoDbError,
-  listingJoiSchema,
   handleJoiValidationError,
-  SchemaValidationError,
 };
